completed inital layer break up
