+++ torchquantum/layer/__init__.py
@@ -25,3 +25,4 @@ SOFTWARE.
 from .layers import *
 from .nlocal import *
 from .general import *
+from .entanglement import *
