+++ torchquantum/layer/layers/u3_layer.py
@@ -1,0 +1,118 @@
+"""
+MIT License
+
+Copyright (c) 2020-present TorchQuantum Authors
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+import torch
+import torch.nn as nn
+import torchquantum as tq
+import torchquantum.functional as tqf
+import numpy as np
+
+from typing import Iterable
+from torchquantum.plugin.qiskit import QISKIT_INCOMPATIBLE_FUNC_NAMES
+from torchpack.utils.logging import logger
+
+from .layers import LayerTemplate0
+
+
+class U3CU3Layer0(LayerTemplate0):
+    """
+    Layer template with U3 and CU3 blocks.
+
+    This layer template consists of U3 and CU3 blocks repeated for the specified number of blocks.
+
+    Args:
+        arch (dict, optional): The architecture configuration for the layer. Defaults to None.
+
+    Attributes:
+        n_wires (int): The number of wires in the layer.
+        arch (dict): The architecture configuration for the layer.
+        n_blocks (int): The number of blocks in the layer.
+        layers_all (tq.QuantumModuleList): The list of layers in the template.
+
+    Methods:
+        build_layers: Builds the U3 and CU3 layers for the template.
+        forward: Applies the quantum layer to the given quantum device.
+
+    """
+
+    name = "u3cu3_0"
+
+    def build_layers(self):
+        layers_all = tq.QuantumModuleList()
+        for k in range(self.arch["n_blocks"]):
+            layers_all.append(
+                Op1QAllLayer(
+                    op=tq.U3, n_wires=self.n_wires, has_params=True, trainable=True
+                )
+            )
+            layers_all.append(
+                Op2QAllLayer(
+                    op=tq.CU3,
+                    n_wires=self.n_wires,
+                    has_params=True,
+                    trainable=True,
+                    jump=1,
+                    circular=True,
+                )
+            )
+        return layers_all
+
+
+class CU3Layer0(LayerTemplate0):
+    """
+    Layer template with CU3 blocks.
+
+    This layer template consists of CU3 blocks repeated for the specified number of blocks.
+
+    Args:
+        arch (dict, optional): The architecture configuration for the layer. Defaults to None.
+
+    Attributes:
+        n_wires (int): The number of wires in the layer.
+        arch (dict): The architecture configuration for the layer.
+        n_blocks (int): The number of blocks in the layer.
+        layers_all (tq.QuantumModuleList): The list of layers in the template.
+
+    Methods:
+        build_layers: Builds the CU3 layers for the template.
+        forward: Applies the quantum layer to the given quantum device.
+
+    """
+
+    name = "cu3_0"
+
+    def build_layers(self):
+        layers_all = tq.QuantumModuleList()
+        for k in range(self.arch["n_blocks"]):
+            layers_all.append(
+                Op2QAllLayer(
+                    op=tq.CU3,
+                    n_wires=self.n_wires,
+                    has_params=True,
+                    trainable=True,
+                    jump=1,
+                    circular=False,
+                )
+            )
+        return layers_all
